requirements.txt
praw==7.7.1
anthropic==0.25.0
twilio==8.13.0
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.31.0
